texk/dvipdfm-x: Minor bug fix and cleanup
git-svn-id: svn://tug.org/texlive/trunk/Build/source@35246 c570f23f-e606-0410-a88d-b1316a301751

+++ texk/dvipdfm-x/bmpimage.c
@@ -99,7 +99,7 @@ get_density (double *xdensity, double *ydensity, struct hdr_info *hdr)
 }
 
 int
-bmp_get_bbox (FILE *fp, long *width, long *height,
+bmp_get_bbox (FILE *fp, int *width, int *height,
               double *xdensity, double *ydensity)
 {
   int r;
